Move stretch_key to shareable_key.rs
